requirements.txt
feedparser>=6.0
requests>=2.31
tqdm>=4.66
google-api-python-client>=2.100
google-auth>=2.23
beautifulsoup4>=4.12
